Refactor: Remove obsolete test files and project references

+++ dtc.API/Controllers/TermController.cs
@@ -3,7 +3,7 @@ using dtc.Application.Features.Training.Interfaces;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using System;
-using System.Security.Claims;
+using System.Linq;
 using System.Threading.Tasks;
 
 namespace dtc.API.Controllers
@@ -22,6 +22,11 @@ namespace dtc.API.Controllers
         public async Task<IActionResult> CreateTerm([FromBody] CreateTermRequestDto request)
         {
             var adminId = await GetInternalUserIdAsync();
+            if (!await CanAccessCourseAsync(request.CourseId))
+            {
+                return Fail("You do not have permission to create a term for this course.");
+            }
+
             try
             {
                 var response = await _termService.CreateTermAsync(request, adminId);
@@ -38,6 +43,11 @@ namespace dtc.API.Controllers
         public async Task<IActionResult> UpdateTerm(Guid id, [FromBody] UpdateTermRequestDto request)
         {
             var adminId = await GetInternalUserIdAsync();
+            if (!await CanAccessTermAsync(id))
+            {
+                return Fail("You do not have permission to access this term.");
+            }
+
             try
             {
                 var response = await _termService.UpdateTermAsync(id, request, adminId);
@@ -54,6 +64,11 @@ namespace dtc.API.Controllers
         public async Task<IActionResult> DeleteTerm(Guid id)
         {
             var adminId = await GetInternalUserIdAsync();
+            if (!await CanAccessTermAsync(id))
+            {
+                return Fail("You do not have permission to access this term.");
+            }
+
             try
             {
                 await _termService.DeleteTermAsync(id, adminId);
@@ -66,15 +81,28 @@ namespace dtc.API.Controllers
         }
 
         [HttpGet]
+        [Authorize(Roles = "Admin,TrainingManager,EnrollmentManager")]
         public async Task<IActionResult> GetAllTerms()
         {
             var response = await _termService.GetAllTermsAsync();
+            var managedCenterId = await GetManagedCenterIdAsync();
+            if (managedCenterId.HasValue)
+            {
+                response = response.Where(term => term.CenterId == managedCenterId.Value);
+            }
+
             return Ok(response);
         }
 
         [HttpGet("{id}")]
+        [Authorize(Roles = "Admin,TrainingManager,EnrollmentManager")]
         public async Task<IActionResult> GetTermDetail(Guid id)
         {
+            if (!await CanAccessTermAsync(id))
+            {
+                return Fail("You do not have permission to access this term.");
+            }
+
             try
             {
                 var response = await _termService.GetTermDetailAsync(id);
